@@ -1,7 +1,10 @@
-﻿using Microsoft.AspNetCore;
+﻿using Microsoft.Extensions.DependencyInjection;
+using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Extensions;
+using Microsoft.Extensions.Hosting;
+using BackEnd.Services.Interfaces;
 
 namespace BackEnd
 {
@@ -9,16 +12,17 @@ namespace BackEnd
     {
         public static void Main(string[] args)
         {
-            BuildWebHost(args).Run();
+            CreateHostBuilder(args).Build().Run();
         }
-        public static IWebHost BuildWebHost(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseConfigFile("appsettings.Secret.json")
-                .UseConfigFile("build.json", true)
-                .ConfigureAppConfiguration(config => config
-                                           .AddCommandLine(args)
-                                           .AddEnvironmentVariables())
-                .UseStartup<Startup>()
-                .Build();
+
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.ConfigureKestrel(options => options.AllowSynchronousIO = true);//TODO disallow
+                    webBuilder.UseStartup<Startup>();
+                })
+                .UseConfigFile("appsettings.Secret.json", optional: true)
+                .UseConfigFile("build.json", true);
     }
 }
