@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Threading.Tasks;
-using WebApp.Configure.Models.Configure.Interfaces;
 using Microsoft.Extensions.Options;
 using BackEnd.Models.Settings;
 using Models.People;
@@ -10,6 +9,10 @@ using Newtonsoft.Json;
 using Models.People.Roles;
 using System.Linq;
 using BackEnd.DataBase;
+using RTUITLab.AspNetCore.Configure.Configure.Interfaces;
+using System.Threading;
+using Models.People.UserProperties;
+using Microsoft.EntityFrameworkCore;
 
 namespace BackEnd.Services.ConfigureServices
 {
@@ -35,10 +38,11 @@ namespace BackEnd.Services.ConfigureServices
             this.logger = logger;
         }
 
-        public async Task Configure()
+        public async Task Configure(CancellationToken cancellationToken)
         {
             if (options.Users?.Any() == true)
                 await CreateUsers();
+            await CreateUserPropertyNames();
             await CreateRoles();
             if (options.WantedRoles?.Any() == true)
                 await ApplyRoles();
@@ -62,6 +66,27 @@ namespace BackEnd.Services.ConfigureServices
             }
         }
 
+        private async Task CreateUserPropertyNames()
+        {
+            foreach (var userPropertyName in Enum.GetValues(typeof(UserPropertyNames)).Cast<UserPropertyNames>())
+            {
+                var internalName = userPropertyName.ToString();
+                var existing = await dbContext.UserPropertyTypes.FirstOrDefaultAsync(upt => upt.InternalName == internalName);
+                if (existing == null)
+                {
+                    var newType = new UserPropertyType
+                    {
+                        DefaultStatus = UserPropertyStatus.NotConfirmed,
+                        InternalName = internalName,
+                        PublicName = internalName,
+                    };
+                    dbContext.UserPropertyTypes.Add(newType);
+                    var saved = await dbContext.SaveChangesAsync();
+                    logger.LogInformation($"Added user property type {internalName}, saved: {saved}");
+                }
+            }
+        }
+
         private async Task ApplyRoles()
         {
             foreach (var wantPair in options.WantedRoles)
@@ -69,7 +94,10 @@ namespace BackEnd.Services.ConfigureServices
                 var targetUser = await userManager.FindByEmailAsync(wantPair.Email);
                 var targetRole = await roleManager.FindByNameAsync(wantPair.RoleName);
                 if (targetUser == null || targetRole == null)
-                    throw new Exception($"Can't find user {wantPair.Email} or role {wantPair.RoleName}");
+                {
+                    logger.LogWarning($"Can't find user \"{wantPair.Email}\" or role \"{wantPair.RoleName}\"");
+                    continue;
+                }
                 if (await userManager.IsInRoleAsync(targetUser, targetRole.Name)) continue;
                 var result = await userManager.AddToRoleAsync(targetUser, targetRole.Name);
                 logger.LogInformation(JsonConvert.SerializeObject(result));
