@@ -40,14 +40,16 @@ namespace IdentityServer
             else
             {
                 services.AddDbContext<DataBaseContext>(options =>
-                    options.UseSqlServer(Configuration.GetConnectionString("DefaultConnection")));
+                    options.UseNpgsql(Configuration.GetConnectionString("DefaultConnection"),
+                        builder => builder.MigrationsAssembly(nameof(BackEnd.DataBase))));
             }
 
             services.AddIdentity<User, Role>()
                 .AddEntityFrameworkStores<DataBaseContext>()
                 .AddDefaultTokenProviders();
 
-            services.AddMvc().SetCompatibilityVersion(Microsoft.AspNetCore.Mvc.CompatibilityVersion.Version_2_1);
+            services.AddMvc(options => options.EnableEndpointRouting = false)
+                .SetCompatibilityVersion(Microsoft.AspNetCore.Mvc.CompatibilityVersion.Version_3_0);
             var migrationsAssembly = typeof(Startup).GetTypeInfo().Assembly.GetName().Name;
 
 
@@ -73,9 +75,10 @@ namespace IdentityServer
             }
             else
             {
-                builder.AddSigningCredential(new X509Certificate2(
+                using var certificate = new X509Certificate2(
                     Configuration.GetValue<string>("ISKeyPath"),
-                    Configuration.GetValue<string>("ISKeyPassword")));
+                    Configuration.GetValue<string>("ISKeyPassword"));
+                builder.AddSigningCredential(certificate);
             }
 
             var corsOrigins = Configuration.GetSection("CORS:Origins").AsEnumerable().Skip(1).Select(kvp => kvp.Value).ToArray();
@@ -119,7 +122,6 @@ namespace IdentityServer
             if (Environment.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
-                app.UseDatabaseErrorPage();
             }
             else
             {
